+++ backend/src/main/java/com/kazmierczak/daniel/car_auction_platform/service/UserServiceImpl.java
@@ -1,11 +1,12 @@
 package com.kazmierczak.daniel.car_auction_platform.service;
 
+import com.kazmierczak.daniel.car_auction_platform.exception.EmailAlreadyTakenException;
 import com.kazmierczak.daniel.car_auction_platform.exception.ResourceNotFoundException;
 import com.kazmierczak.daniel.car_auction_platform.repository.UserRepository;
 import com.kazmierczak.daniel.car_auction_platform.dto.UserDto;
 import com.kazmierczak.daniel.car_auction_platform.entity.User;
 import com.kazmierczak.daniel.car_auction_platform.mapper.UserMapper;
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -13,14 +14,11 @@ import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
+@RequiredArgsConstructor
 public class UserServiceImpl implements UserService {
 
     private final UserRepository userRepository;
 
-    @Autowired
-    public UserServiceImpl(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
 
     @Override
     public List<UserDto> getAll() {
@@ -42,12 +40,18 @@ public class UserServiceImpl implements UserService {
     @Override
     public UserDto saveUser(UserDto userDto) {
         User user = UserMapper.toEntity(userDto);
+        if (user.getId() == null && userRepository.findByEmail(user.getEmail()).isPresent()) {
+            throw new EmailAlreadyTakenException("User with email " + user.getEmail() + " already exists");
+        }
         User savedUser = userRepository.save(user);
         return UserMapper.toDto(savedUser);
     }
 
     @Override
     public void deleteById(Long id) {
+        if(!userRepository.existsById(id)){
+            throw new ResourceNotFoundException("Cannot delete. User with id " + id + " not found.");
+        }
         userRepository.deleteById(id);
     }
 }
